+++ parivesh/src/main/java/com/backend/repository/postgres/ActivityRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import com.backend.model.activities;
 
 
-public interface ActivityRepository extends JpaRepository<activities, Long>{
+public interface ActivityRepository extends JpaRepository<activities, Integer>{
 
 	
 	@Query("SELECT new activities(a.id, a.name) from activities a where a.is_active='true' and a.is_deleted='false'")
